feat: merge fingerprint features
