+++ build.js
@@ -36,7 +36,7 @@ esbuild.build({
     target: ["esnext"],
 }).catch(() => process.exit(1));
 
-const entryPoints = ["src/domain/index.ts", "src/domain/card/index.ts", "src/domain/fingerprint/index.ts", "src/domain/iframe/utils.ts"]
+const entryPoints = ["src/domain/index.ts", "src/domain/card/index.ts", "src/domain/antifraud/fingerprint/index.ts", "src/domain/iframe/utils.ts"]
 
 // Builda os pacotes separados com esm (para ser usado com 'import')
 esbuild.build({
@@ -70,6 +70,16 @@ esbuild.build({
     minify: false,
     format: "cjs",
     outfile: "dist/cjs/domain/card/index.cjs.js",
+    target: ["esnext"],
+    define: env
+}).catch(() => process.exit(1));
+
+esbuild.build({
+    entryPoints: ["src/domain/antifraud/fingerprint/index.ts"],
+    bundle: true,
+    minify: false,
+    format: "cjs",
+    outfile: "dist/cjs/domain/antifraud/fingerprint/index.cjs.js",
     target: ["esnext"],
     define: env
 }).catch(() => process.exit(1));
