+++ src/domain/antifraud/types.ts
@@ -1,0 +1,1 @@
+export type AntifraudService = "NETHONE" | "OSCILAR" | "BARTE";
